add securing things to view

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, reverse, redirect, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -69,9 +70,12 @@ def product_detail(request, product_id):
     
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     '''add a product to th store'''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you have no access to change products!')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -89,8 +93,12 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
+    ''' Edit a product details'''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you have no access to change products!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -112,8 +120,12 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
+    ''' delete a product srom the shop '''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you have no access to change products!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted')
